fix: Make the crate compatible with recent nightlies
