+++ tests/main.glu
@@ -1,0 +1,2 @@
+let x = 1
+x
